refactor: separate search result
